% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_clone_matrix}
\alias{as_clone_matrix}
\title{Builds a adjacency matrix of clone and cells}
\usage{
as_clone_matrix(clones, cells, cell_order = NULL)
}
\arguments{
\item{clones}{vector of clone ids}

\item{cells}{vector or list of cell ids with the same length as \code{clones}}

\item{cell_order}{ordered vector of cell ids indicating their row position.
When \code{NULL} it uses the sorted order of elements,}
}
\value{
a sparse adjacency matrix indicating which cell belongs to which
clone.
}
\description{
Given the cell composition of each clone, it builds and returns a binary
adjacency sparse matrix where each clone corresponds to a column, each
cell to a row and a 1 value indicates that cell belongs to the clone.
}
