% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_nb.R
\name{clone_nb}
\alias{clone_nb}
\title{Perturbed clone and gene differential expression}
\usage{
clone_nb(
  interactions,
  count,
  clone,
  ...,
  min_x = 1,
  min_n = 2,
  theta_min_mu = 0.05,
  theta_n = 2000,
  gene_col = "gene",
  clone_col = "clone"
)
}
\arguments{
\item{interactions}{table with clone and gene indexes pairs to measure
differential gene expression.}

\item{count}{count matrix with cells as columns and features as rows.}

\item{clone}{logical or adjacency matrix indicating the perturbed cells
(rows) belonging to each clone (columns).}

\item{...}{currently ignored}

\item{min_x}{minimal average expression of perturbed and non-perturbed cells
required.}

\item{min_n}{minimal number of perturbed cells required.}

\item{theta_min_mu}{minimal overall average expression required to estimate
gene \code{theta} and run test.}

\item{theta_n}{number of genes sampled to preliminary \code{theta} estimation.}

\item{gene_col}{name or index of feature column in \code{interactions}}

\item{clone_col}{name or index of clone column in \code{interactions}}
}
\value{
object of the type as \code{interactions} with the following columns added:
\itemize{
\item \code{n0} and \code{n1}: number of non-perturbed and perturbed cells
\item \code{x0} and \code{x1}: number of non-perturbed and perturbed cells average count
\item \code{mu}: overall average expression
\item \code{theta}: negative binomial overdispersion
\item \code{xb}: perturbed cells' estimated average count
\item \code{z}: perturbed cells' standardize z-score effect
\item \code{lfc}: perturbed cells' log2 fold-change effect
\item \code{pvalue}
\item \code{qvalue}
}
}
\description{
Given a table of clone and gene interactions pairs, UMI count matrix and
clone adjacency matrix, it models each interaction gene expression (\code{Y})
with a negative binomial regression in function of cell total count (\code{D})
and a clone indicator variable (\code{X}) as indicated by the model below:
}
\details{
\deqn{Y ~ NB(xb, theta)}
\deqn{log(xb) = \beta_0 + \beta_d * log(D) + \beta_x * X}
\deqn{theta ~ \mu}

Each gene overdispersion (\code{theta}) is estimated in two steps. First, it fits
a Poisson regression as function of log total cell count and uses maximum
likelihood estimator to estimate the raw \code{theta} for a subset of genes.
Next it regularize and expands \code{theta} estimates as function of overall
average count.
}
