% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cellranger.R
\name{read_cellranger}
\alias{read_cellranger}
\title{Load data from cellranger project}
\usage{
read_cellranger(data_dir)
}
\arguments{
\item{data_dir}{Directory containing matrix.mtx, features.tsv, and
barcodes.tsv file produced by 10X run.}
}
\value{
a sparse matrix containing the expression data.
}
\description{
Enables easy loading of 10x single-cell experiment sparse matrix data.
Currently compatible only with Cell Ranger >= v3.0
}
