% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_simulated}
\alias{sc_simulated}
\title{Simulated scRNA-seq count matrix (1500 x 500)}
\format{
A sparse count matrix with 1500 rows (features) and 500 columns
(cells).
}
\usage{
sc_simulated
}
\description{
The matrix was generated using \code{sim_count()} for a total cell depth
following a log-normal distribution (\eqn{ln(D) ~ N(9, .5)}), a gene
expression rate ranging from 0.0001 to 0.01 in a logarithm scale
and overdispersion (\code{theta}) of 80.
}
\keyword{datasets}
