% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_log.R
\name{seq_log}
\alias{seq_log}
\title{Sequence generation in exponential scale}
\usage{
seq_log(from, to, length_out, base = 10L)
}
\arguments{
\item{from, to}{starting (minimal) and end (maximal) values fo the sequence}

\item{length_out}{desired length of the sequence}

\item{base}{logarithm scale base for which the data is spaced}
}
\description{
Generates a numeric sequence with a specific length spaced equally in an
logarithm scale.
}
