% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_experiment.R
\name{sim_experiment}
\alias{sim_experiment}
\title{Simulate a clone perturbation experiment}
\usage{
sim_experiment(
  reference,
  log2_fc = 0L,
  n_clones = 10L,
  n_genes = NULL,
  min_n = 2L,
  max_n = 20L,
  theta = 100L
)
}
\arguments{
\item{reference}{reference UMI count matrix}

\item{log2_fc}{log2 fold-change to apply to each interaction of a clone}

\item{n_clones}{number of clones to generate}

\item{n_genes}{number of genes to generate, if \code{NULL} it will generate
the same number of genes as \code{reference}}

\item{min_n}{minimum clone size}

\item{max_n}{maximum clone size}

\item{theta}{genes' negative binomial dispersion}
}
\value{
list containing table of gene and clone \code{interactions}, \code{clone}
matrix and \code{count} matrix.
}
\description{
Given a reference single-cell UMI count matrix, it generates a random set of
clones, a random count matrix, a table of clone and gene interactions and
the applied log2 fold-change.
}
\details{
First, it produces a clone matrix using \code{sim_clone_range} to produce clones
size between \code{min_n} and \code{max_n}. Next, for each clone it generates a gene
interaction for each \code{log2_fc}. Finally, it computes the intended log2
fold-change effect and generates a random count matrix.
}
