% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_clone.R
\name{simulate_clone}
\alias{simulate_clone}
\alias{sim_clone}
\alias{sim_clone_range}
\title{Simulate a random clonal assignment}
\usage{
sim_clone(size, ncells)

sim_clone_range(n, ncells, min = 2L, max = 20L)
}
\arguments{
\item{size}{number of cells to be assigned to each clone}

\item{ncells}{total number of cells available to assign}

\item{n}{total number of clones be generated}

\item{min}{minimum number of cells perturbed by clone}

\item{max}{maximum number of cells perturbed by clone}
}
\value{
a list of cells per clone.
}
\description{
It generates a random clone assignment list of specific sizes.
}
