% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_nb.R
\name{clone_nb}
\alias{clone_nb}
\title{Perturbed clone differential expression}
\usage{
clone_nb(
  pertubed,
  count,
  clone,
  ...,
  min_x = 1,
  min_n = 2,
  theta_min_mu = 0.05,
  theta_n = 2000,
  gene_col = "gene",
  clone_col = "clone"
)
}
\arguments{
\item{pertubed}{A data.frame like structure describing clone and gene pairs
to model expression and measure perturbation.}

\item{count}{A \code{j} x \code{z} logical matrix or sparseMatrix indicating if the
\emph{jth} cell belongs or not to \emph{zth} clone.}

\item{clone}{A \code{i} x \code{j} count matrix or sparseMatrix indicating if the UMI
expression of \emph{ith} gene for the  \emph{jth} cell.}

\item{...}{Currently ignored}

\item{min_x}{Minimal average expression within or outside the pertubed clone
(\code{x1} and \code{x0}, respectively) required to test effect.}

\item{min_n}{Minimal number of cells required to test a clone.}

\item{theta_min_mu}{Minimal \code{mu} (overall average expression) required to
estimated gene \code{theta} and to be tested.}

\item{theta_n}{Number of genes sampled to estimate \code{theta}.}

\item{gene_col}{Name or index of \code{gene} column in \code{pertubed}.}

\item{clone_col}{Name or index of \code{clone} column in \code{pertubed}.}
}
\value{
A object of the same type as \code{pertubed} added the following columns:
}
\description{
It takes scRNA-seq UMI count matrix, clone assignment matrix and compares
the expression of a given gene within and outside a given clone.
}
\details{
The UMI counts are modeled by a Negative Binomial regression
(\eqn{Y_ij ~ NB(\mu_ij, \theta_i)}), where \eqn{\theta_i} is the gene \eqn{i}
dispersion and \eqn{\mu_ij} is gene \eqn{i} and cell \eqn{j} expected UMI
count estimated as follow:
\deqn{log(\mu_ij) = \beta_0 + \beta_1 \times log(dp_j) + \beta_2 x_ij}
where \eqn{dp_j} is cell \eqn{j} total UMI depth and \eqn{x_ij} is a
\emph{i.i.d.} measure if cell \eqn{j} belonged to a clone perturbing gene
\eqn{i} expression.
}
