% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_clone.R
\name{simulate_clone}
\alias{simulate_clone}
\alias{sim_clone}
\alias{sim_clone_range}
\title{Simulate a random clone matrix}
\usage{
sim_clone(size, n_cells)

sim_clone_range(n_clones, n_cells, min = 2L, max = 20L)
}
\arguments{
\item{size}{number of cells to be assigned to each clone}

\item{n_cells}{total number of cells}

\item{n_clones}{total number of clones be generated}

\item{min}{minimum number of cells perturbed by clone}

\item{max}{maximum number of cells perturbed by clone}
}
\value{
adjacency matrix indicating which cells (rows) belong to each clone
(column).
}
\description{
Given a vector of clone sizes (\code{perturbed}) and the total number of cells
(\code{n_cells}) and generates a random adjacency matrix with \code{n_cells} rows and
\code{length(perturbed)} columns.
}
