% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_count.R
\name{simulate_count}
\alias{simulate_count}
\alias{sim_count}
\alias{sim_count_from_data}
\title{Simulates a single-cell experiment UMI count matrix}
\usage{
sim_count(cell_depth, gene_rate, log2_fc = 0L, theta = 100L)

sim_count_from_data(reference, log2_fc = 0L, theta = 100L, n_genes = NULL)
}
\arguments{
\item{cell_depth}{total UMI count per cell}

\item{gene_rate}{gene rate (gene total UMI divided by matrix total UMI)}

\item{log2_fc}{log2 fold-change effect among the cells}

\item{theta}{genes negative binomial dispersion parameter}

\item{reference}{count matrix to be used as reference}

\item{n_genes}{number of genes to simulate. If set to \code{NULL} it
produces the gene rate observed in \code{reference}, otherwise if a postive
integer scalar is given, it simulates \code{n_genes} with a gene rate spanning
from lowest to highest observed in a logarithm range observed.}
}
\value{
random count matrix
}
\description{
It simulates a random single-cell experiment UMI count matrix based on our
model assumptions given the total cell count (\code{cell_depth}) and gene rate
(genes total count divided by total count; \code{gene_rate}).
You can further specify the model, by giving a log2 fold-change effect to
apply by cell or by cell and gene (\code{log2_fc}) and each gene negative
binomial model dispersion (\code{theta}).
}
